# openapi.yaml
openapi: 3.1.0
info:
  title: NovaMember API
  version: 1.0.0
  description: |
    Shopify + Recharge membership management platform with AI-powered engagement scoring.
    All authenticated endpoints require a Bearer JWT obtained from `/api/auth/login` or `/api/checkout`.
  contact:
    name: NovaMember Support
    email: support@novamember.io
  license:
    name: Proprietary
servers:
  - url: https://novamember.io
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/webhooks/recharge:
    post:
      tags: [Webhooks]
      summary: Recharge webhook receiver
      description: |
        HMAC-verified. Handles: subscription/activated, subscription/cancelled,
        charge/paid, charge/failed. Fully idempotent — replayed events are ignored.
      security:
        - rechargeHmac: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Received (always 200 to prevent Recharge retry storms)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT from `/api/auth/login` or `/api/checkout`
    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: Vercel cron secret from `CRON_SECRET` env var
    rechargeHmac:
      type: apiKey
      in: header
      name: X-Recharge-Hmac-Sha256
    shopifyHmac:
      type: apiKey
      in: header
      name: X-Shopify-Hmac-Sha256
...